+++ CHANGELOG.md
@@ -13,6 +13,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Fixed
 
+- Toon badge optie bij politie menu verwijderd (vervangen door item)
+
 ## [1.3.0]
 
 ### Added