+++ CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Wat metadata van de flyer gehide
 - Flyer links worden nu treffelijk gecontroleerd of ze wel naar een image wijzen
+- Stash grootte van house wordt nu bepaald door house type
 
 ### Fixed